+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Fixes
 
 - Can't drop item with NeoForge
+- Every entity on a playback had a visible name (like items dropped)